+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Improve MapperUtils.convertValue 
-- Add MapperUtils.readValue for java class 
+- Add MapperUtils.readValue for java class
+- Add request executor for requests with extremum Response model
 
 ## [3.0.0] - 2023-03-24